Fix Ohanakapa search bug
After upgrading to 3.1.2, the Ohanakapa.search method broke, but the tests didn't catch it because they weren't using the same middleware as other environments. I don't recall why I had excluded it from tests back in 2015.

In this PR, I removed the exclusion, which then caused the tests to fail as they should. I also added a location spec for extra safety.

The stack trace showed the error coming from the faraday-http-cache gem, and upgrading it solved the issue.