Update Ruby from 3.1.4 to 3.2.3
This resulted in an exception from the compass gem, so I looked into getting rid of the compass gem dependency, and got it working. For a few things related to typography, I had to copy over the functions from the compass repo. Otherwise, I replaced the compass functions with plain CSS.