Replace webpacker with jsbundling-rails
Reference: https://github.com/rails/jsbundling-rails/blob/main/docs/switch_from_webpacker.md

Due to the asset_path references in markers.js.erb, and because jsbundling-rails
adds a "build" entry in package.json, we can no longer use the Node buildpack on Heroku.

If the Node buildpack is present, and comes before the Ruby buildpack, it will try to run the `build` script, but in order to compile/transpile the js.erb, the `rails-erb-loader` package requires Ruby to be present, and so it fails because Heroku hasn't installed Ruby yet.

With Webpacker, there was no `build` script, so we didn't run into this issue. The JS compilation happened as part of `rake assets:precompile`.

Maybe there's a way to remove the build script from package.json and add our own rake task that runs `yarn build` as part of asset precompilation?

For now, the solution was to remove the Node buildpack. The downside is that you can't specify the version of Node or Yarn. Heroku will decide which version to use. That's why we bumped the Node version from 18 to 20.