Update Ruby from 3.2.4 to 3.2.5
This initially resulted in an error when installing rbtree, which is a dependency of sorted_set, but it doesn't look like sorted_set is used, so I removed it.

It looks like the sorted_set gem was added to resolve issues with compass and compass-rails: https://github.com/smcgov/SMC-Connect/commit/01390e29bbe31489cff10e22a5e7b17a85edf908

But those gems were removed a while back, so we don't need sorted_set anymore.